+++ app/controllers/home_controller.rb
@@ -1,7 +1,5 @@
 class HomeController < ApplicationController
   def show
-    render inertia: "Home/Show", props: {
-      appName: "Routeprint"
-    }
+    render inertia: "Home/Show"
   end
 end